refactor(host-infra-system): split git adapter into focused modules
